requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.0
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.0
langchain-groq==0.1.5
ollama==0.2.1
chromadb==0.5.3
sentence-transformers==3.0.1
tiktoken==0.7.0
json-repair==0.25.3
streamlit==1.35.0
streamlit-extras==0.4.0
streamlit-lottie==0.0.5
requests==2.31.0
duckduckgo-search==6.2.12
beautifulsoup4==4.12.3
pypdf==4.2.0
fal-client==0.7.0
Pillow==10.3.0
imageio==2.34.2
imageio-ffmpeg==0.5.1
numpy==1.26.4
slack-sdk==3.31.0
sqlalchemy==2.0.31
diffusers==0.27.2
transformers==4.41.2
accelerate==0.31.0
bitsandbytes==0.43.1
xformers==0.0.26.post1
torch==2.2.0+cu121
torchvision==0.17.0+cu121
torchaudio==2.2.0+cu121
